Split up ResultPrinter and SummaryPrinter and reused code in DetailedResultPrinter
